[should-not-exist-rule] rename class and move functions to extractor trait
