+++ docs/documentation/assertions.md
@@ -48,7 +48,7 @@ It asserts that the selected classes **do not depend** on the target classes.
 It asserts that the selected classes **do not use the constructor** of the target classes.
 
 ## shouldNotExist()
-It asserts that the selected classes **should not exist**.
+It asserts that the selected classes **do not exist**.
 
 ## shouldApplyAttribute()
 It asserts that the selected classes **apply** the target attributes.
